legacy/requirements.txt
# Interface e Sistema
PyQt5>=5.15.9
Pillow>=10.0.0
keyboard>=0.13.5

# OCR e Processamento de Imagem
pytesseract>=0.3.10
opencv-python>=4.8.0.76

# HTTP e API
requests>=2.31.0

# Logging avançado
colorlog>=6.7.0

# Segurança e Criptografia
cryptography>=41.0.0
python-dotenv>=1.0.0
bcrypt>=4.0.1

# Rate Limiting e Controle de Acesso
limits>=3.5.0
slowapi>=0.1.9

# Análise e Validação
pydantic>=2.0.0
validators>=0.20.0

# Performance e Monitoramento
psutil>=5.9.0
memory-profiler>=0.60.0

# Qualidade de Código (desenvolvimento)
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
pytest>=7.4.0
pytest-cov>=4.1.0
bandit>=1.7.5

# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database & Auth
supabase==2.0.2
psycopg2-binary==2.9.9

# HTTP & Async
httpx==0.25.2
aiofiles==23.2.1

# Environment & Config
pydantic==2.5.0
pydantic-settings==2.1.0

# AI & ML
openai==1.3.7
anthropic==0.7.8
groq==0.4.1

# Utils
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development
pytest-asyncio==0.21.1
isort==5.12.0 